// src/ext/annonars-api/src/lib/types.gen.ts
export type GenesClingenDosageScore =
  | 'SufficientEvidenceAvailable'
  | 'SomeEvidenceAvailable'
  | 'LittleEvidence'
  | 'NoEvidenceAvailable'
  | 'Recessive'
  | 'Unlikely'